requirements.txt
openpyxl
pywaclient~=0.6.5
requests
simple-elastic
PyPDF2
python-dotenv
pyswagger